td error related.

+++ bridger/builder_trainer.py
@@ -273,7 +273,7 @@ class BridgeBuilderModel(pl.LightningModule):
                 _, _, start_state, _, _, _, _ = next(self.memories)
                 if self.hparams.debug:
                     self._state_visit_logger.log_occurrence(
-                        batch_idx=batch_idx, object=torch.Tensor(start_state)
+                        batch_idx=batch_idx, object=torch.Tensor(start_state).double()
                     )
 
     def _memory_generator(
@@ -533,7 +533,7 @@ class BridgeBuilderModel(pl.LightningModule):
                     loss=loss,
                 ),
             )
-
+            
             for state, action, td_error in zip(
                 states, actions.tolist(), td_errors.tolist()
             ):
